add basic syntax error checking, textfileposition changes
if a matcher has multiple things it needs to match, only the first one needs to return null since when one part of a languageelement is matched, it must be that element (pretty sure). this means further checks can throw errors if they fail, since something is (probably) certainly wrong.

also added EOF type to TextFilePosition and moved it to templater.compiler